reorganize files into extensions directory
