+++ lib/thegarage/gitx/extensions/string.rb
