+++ lib/thegarage/gitx/extensions/thor.rb
